UNMERGED: drm: rockchip: add rk3288 dmc support
At first I want to let it share code with 3399 code,
But they are really different and we might need to change lots code.

Since the 3288 DMC is a trick way and upstream won't merge it,
it would be better to just put those code in macro.Although ugly,
but also clearly.

Change-Id: I013efbe8a42b028319e1c6cfc37f5dcbc6e2cb01
Signed-off-by: Jacob Chen <jacob2.chen@rock-chips.com>